animate when opening a stream
- Animation when starting a stream
- Errors now shows to the user inside a message box
- Calendar now dissapear by clicking beside it
- Downloader class now has a parameter to not download a file if the
delay is not over (delay is in month, day or hour)
- Default download path is now to NHLGames directory, if app has no
access it will try to write in appdata, then in temp (It's to let the
user know we are downloading files on his computer, so he could remove
it he want)
- Choosen download path now shows in console
- Not attempting to download version.txt if the file is too recent.
Attemps are set every month at the moment.
- Not attempting to download games.txt if the file is too recent.
Attemps are set every day at the moment
- Clean .json files in download repertory at every launch at the moment.
- Removed old code that was not used (so look at the changes, like
fileaccess.vb for detecting an .exe)
- hide url that livestreamer was showing in the console after not being
able to open a .m3u8 link (not sure, if this is supposed to be hidden)
- Links to download version.txt and changelog.txt are now https, i
removed the 'if' condition because it was downloading more than one time
the same file.
- The layout panel does not resize the form anymore, when not in
fullscreen the height was bigger than the screen size. Instead the
layout is inside another layout. the parent layout is fixed and anchor
to the form, so a vertical scrollbar appears if too many games are
found.
- Progress bar should now loads porperly